RAG/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "json-repair>=0.46.0",
    "langchain-google-genai>=2.1.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.3.0",
    "langchain-ollama>=0.3.3",
    "lxml>=5.4.0",
    "neo4j>=5.28.1",
    "ollama>=0.5.0",
    "pydantic>=2.11.5",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "unstructured[all-docs]>=0.17.2",
    "wikipedia>=1.4.0",
    "yfiles-jupyter-graphs>=1.10.6",
]
